bbs commit

+++ src/main/java/com/sparta/bbs/service/BbsService.java
@@ -3,6 +3,7 @@ package com.sparta.bbs.service;
 
 import com.sparta.bbs.dto.BbsResponseDto;
 import com.sparta.bbs.dto.DeleteBbsDto;
+import com.sparta.bbs.dto.ResponseDto;
 import com.sparta.bbs.entity.Member;
 import com.sparta.bbs.jwt.JwtUtil;
 import com.sparta.bbs.repository.BbsRepository;
@@ -11,6 +12,7 @@ import com.sparta.bbs.entity.Bbs;
 import com.sparta.bbs.repository.MemberRepository;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
@@ -44,7 +46,7 @@ public class BbsService {
     public BbsResponseDto getPost(Long id) {
         Bbs bbs = bbsRepository.findById(id).orElseThrow();
         BbsResponseDto responseDto = new BbsResponseDto(bbs);
-        responseDto.getClass();
+//        responseDto.getClass();
         return responseDto;
     }
 
@@ -70,18 +72,13 @@ public class BbsService {
             //토큰에서 가져온 사용자 정보를 사용하여 DB 조회
             //claims.getSubject(): 넣어두었던 username 가져오기
             //findByUsername()를 사용해서, UserRepository 에서 user 정보를 가져오기
-            Member member = memberRepository.findByUsername(claims.getSubject()).orElseThrow(
-                    //매개변수가 의도치 않는 상황 유발시
-                    () -> new IllegalArgumentException("사용자가 존재하지 않습니다.")
-            );
+            Member member = checkUser(claims);
 
             //요청받은 DTO 로 DB에 저장할 객체 Board board 만들기
-            Bbs bbs = bbsRepository.saveAndFlush(new Bbs(bbsRequestDto, member.getId()));
-            System.out.println(bbs);
-            System.out.println(bbs.getId());
-            System.out.println(bbsRequestDto);
+            Bbs bbs = new Bbs(bbsRequestDto, member.getId(), member.getUsername());
+            Long id = bbsRepository.saveAndFlush(bbs).getId();
             //entity(board) 를 DTO(BoardResponseDto) 로 변환시켜서 리턴(반환)
-            return new BbsResponseDto(bbs);
+            return new BbsResponseDto(checkPost(id));
             //토큰이 null 이라면(Client 에게서 Token 이 넘어오지 않은 경우),
         } else {
 
@@ -97,9 +94,7 @@ public class BbsService {
     @Transactional
     public BbsResponseDto update(Long id, BbsRequestDto bbsRequestDto, HttpServletRequest httpServletRequest) {
         String token = jwtUtil.resolveToken(httpServletRequest);
-
         Claims claims;
-
 
         if (token != null) {
             // validateToken => 토큰 검증
@@ -108,15 +103,13 @@ public class BbsService {
             } else {
                 throw new IllegalArgumentException("Token Error");
             }
-            // 토큰에서 가져온 사용자 정보를 사용하여 DB 조회
-            //claims.getSubject(): 우리가 넣어두었던 username 가져오기
-            //findByUsername()를 사용해서, UserRepository 에서 user 정보를 가져오기
-            Member member = memberRepository.findByUsername(claims.getSubject()).orElseThrow(() -> new IllegalArgumentException("사용자가 존재하지 않습니다."));
+            Member member = checkUser(claims);
+            Bbs bbs = checkPost(id);
 
-            //즉, 현재 로그인한 user 가 선택한 board 가 맞는지 확인
-            Bbs bbs = bbsRepository.findByIdAndUserId(id, member.getId()).orElseThrow(() -> new IllegalArgumentException("게시글이 존재하지않습니다."));
-            bbs.update(bbsRequestDto);
-            return new BbsResponseDto(bbs);
+            if (bbs.getUsername().equals(member.getUsername())) {
+                bbs.update(bbsRequestDto);
+            }
+            return new BbsResponseDto(checkPost(id));
         } else {
             return null;
         }
@@ -124,7 +117,7 @@ public class BbsService {
 
     // 삭제
     @Transactional
-    public DeleteBbsDto delete(Long id, BbsRequestDto bbsRequestDto, HttpServletRequest httpServletRequest) {
+    public ResponseDto delete(Long id, HttpServletRequest httpServletRequest) {
 
         String token = jwtUtil.resolveToken(httpServletRequest);
         //JWT 안에 있는 정보를 담는 Claims 객체
@@ -137,25 +130,29 @@ public class BbsService {
                 throw new IllegalArgumentException("Token Error");
             }
 
-            Member member = memberRepository.findByUsername(claims.getSubject()).orElseThrow(() -> new IllegalArgumentException("사용자가 존재하지않습니다."));
+            Member member = checkUser(claims);
+            Bbs bbs = checkPost(id);
 
-            Bbs bbs = bbsRepository.findByIdAndUserId(id, member.getId()).orElseThrow(() -> new IllegalArgumentException("게시글이 존재하지않습니다."));
-            bbsRepository.delete(bbs);
-
-            return new DeleteBbsDto("게시글 삭제성공", 200);
+            if (bbs.getUsername().equals(member.getUsername())) {
+                bbsRepository.delete(bbs);
+            }
+            return new ResponseDto("삭제 성공", 200);
 
         } else {
-            return new DeleteBbsDto("게시글 작성자만 삭제 가능", HttpStatus.OK.value());
+            return new ResponseDto("게시글 작성자만 삭제 가능", 100);
         }
-
-//        DeleteBbsDto deleteBbsDto = new DeleteBbsDto();
-//        if (bbsRequestDto.getPassword().equals(bbs.getPassword())) {
-//            bbsRepository.deleteById(id);
-//            deleteBbsDto.setMsg("success");
-//        } else {
-//            deleteBbsDto.setMsg("failed");
-//        }
-//        return deleteBbsDto;
-//    }
     }
+
+    private Bbs checkPost(Long id){
+        return  bbsRepository.findById(id).orElseThrow(
+                () -> new IllegalArgumentException("해당 포스트가 없습니다.")
+        );
+    }
+    private Member checkUser(Claims claims) {
+        return memberRepository.findByUsername(claims.getSubject()).orElseThrow(
+                () -> new IllegalArgumentException("사용자가 존재하지 않습니다.")
+        );
+    }
+
+
 }
